requirements.txt
python-dotenv
python-telegram-bot
openai
telegramify_markdown